user profile and Privilege
Modify Connectivity table. Update requirement on user profile and Privilege

+++ L20_RTP/AlarmData.h
@@ -24,7 +24,7 @@ struct ALARM_DATA
 {
 	unsigned int 	AlarmType;
 	unsigned int	AlarmSubType;
-	long			WeldResultID;
+	unsigned int	WeldResultID;
 	unsigned int	WeldRecipeID;
 	unsigned int	UserID;
 };
